Windows Handler is up and running. Next Step is to get drawing with OpenGL

+++ Project Diode/Project Diode/Callable.cpp
@@ -29,7 +29,7 @@
 
 */
 /******************************************************************************/
-Callable::Callable(long id, std::function<void(std::shared_ptr<Message> const&)> fxn)
+Callable::Callable(long long id, std::function<void(std::shared_ptr<Message> const&)> fxn)
 {
   mId = id;
   mFunction = fxn;
